Basic pixel data handler registration
- Client registers observers on startup for each loaded channel

- Fixed a few bugs on server side preventing registration and removal of
  registrations from working properly

- Fixed an issue where the random engine was re-created always in
  creating tokens for the framebuffer observers, resulting in the same
  token values being produced over and over again.

+++ client/plugin/dummy/DummyChannel.h
@@ -6,7 +6,7 @@
  */
 class DummyChannel: public Lichtenstein::Client::Output::IOutputChannel {
     public:
-        DummyChannel(size_t i, size_t _numPixels, size_t _format) : index(i), 
+        DummyChannel(uint32_t i, size_t _numPixels, size_t _format) : index(i), 
             numPixels(_numPixels), format(_format) {}
 
     public:
@@ -21,14 +21,14 @@ class DummyChannel: public Lichtenstein::Client::Output::IOutputChannel {
         }
 
         /// Channel index
-        virtual size_t getChannelIndex() {
+        virtual uint32_t getChannelIndex() {
             return this->index;
         }
 
         virtual int updatePixelData(const void *data, const size_t dataLen);
 
     private:
-        size_t index = 0;
+        uint32_t index = 0;
         size_t numPixels = 0;
         size_t format = 0;
